@@ -1,105 +1,130 @@
-import {SafeAreaView, ScrollView, StyleSheet, View, Text, StatusBar, Button, Platform} from 'react-native';
+import {
+  SafeAreaView,
+  ScrollView,
+  StyleSheet,
+  View,
+  Text,
+  StatusBar,
+  Button,
+  Platform,
+} from 'react-native';
 import DateTimePicker from '@react-native-community/datetimepicker';
-import {Header,Colors} from 'react-native/Libraries/NewAppScreen';
-import React, {Fragment, Component} from 'react';
+import {Header, Colors} from 'react-native/Libraries/NewAppScreen';
+import React, {useState} from 'react';
 import moment from 'moment';
 
-type Props = {};
-export default class App extends Component<Props> {
-  state = {
-    date: new Date(1598051730000),
-    mode: 'date',
-    display: 'default',
-    show: false,
-    interval: undefined,
+const App = () => {
+  const [date, setDate] = useState(new Date(1598051730000));
+  const [mode, setMode] = useState('date');
+  const [show, setShow] = useState(false);
+  const [display, setDisplay] = useState('default');
+  const [interval, setMinInterval] = useState(undefined);
+
+  const onChange = (event, selectedDate) => {
+    const currentDate = selectedDate || date;
+
+    setShow(Platform.OS === 'ios');
+    setDate(currentDate);
   };
 
-  setDate = (event, date) => {
-    date = date || this.state.date;
-
-    this.setState({
-      show: Platform.OS === 'ios',
-      date,
-    });
+  const showMode = currentMode => {
+    setShow(true);
+    setMode(currentMode);
+    setMinInterval(undefined);
   };
 
-  show = (mode, display = 'default', interval = 1) => {
-    this.setState({
-      show: true,
-      mode,
-      display,
-      interval,
-    });
+  const showDatepicker = () => {
+    showMode('date');
+    setDisplay('default');
   };
 
-  datepicker = () => {
-    this.show('date');
+  const showDatepickerSpinner = () => {
+    showMode('date');
+    setDisplay('spinner');
   };
 
-  timepicker = () => {
-    this.show('time');
+  const showTimepicker = () => {
+    showMode('time');
   };
 
-  timepickerSpinner = () => {
-    this.show('time', undefined, 5);
+  const showTimepickerWithInterval = () => {
+    showMode('time');
+    setDisplay('spinner');
+    setMinInterval(5);
   };
 
-  render() {
-    const {show, date, mode, display, interval} = this.state;
-
-    return (
-      <Fragment>
-        <StatusBar barStyle="dark-content"/>
-        <SafeAreaView>
-          <ScrollView
-            contentInsetAdjustmentBehavior="automatic"
-            style={styles.scrollView}>
-            <Header/>
-            {global.HermesInternal == null ? null : (
-              <View style={styles.engine}>
-                <Text style={styles.footer}>Engine: Hermes</Text>
-              </View>
-            )}
-            <View style={styles.body}>
-              <View testID="appRootView" style={styles.container}>
-                <View style={styles.header}>
-                  <Text style={styles.text}>Example DateTime Picker</Text>
-                </View>
-                <View style={styles.button}>
-                  <Button testID="datePickerButton" onPress={this.datepicker} title="Show date picker!"/>
-                </View>
-                <View style={styles.button}>
-                  <Button testID="timePickerButton" onPress={this.timepicker} title="Show time picker (default)!"/>
-                </View>
-                <View style={styles.button}>
-                  <Button testID="timePickerIntervalButton" onPress={this.timepickerSpinner} title="Show time picker (with 5 min interval)!"/>
-                </View>
-                <View style={styles.header}>
-                  <Text testID="dateTimeText" style={styles.dateTimeText}>
-                    {mode === 'time' && moment.utc(date).format('HH:mm')}
-                    {mode === 'date' && moment.utc(date).format('MM/DD/YYYY')}
-                  </Text>
-                </View>
-                {show && (
-                  <DateTimePicker
-                    minuteInterval={interval}
-                    testID="dateTimePicker"
-                    timeZoneOffsetInMinutes={0}
-                    value={date}
-                    mode={mode}
-                    is24Hour={true}
-                    display={display}
-                    onChange={this.setDate}
-                  />
-                )}
-              </View>
+  return (
+    <>
+      <StatusBar barStyle="dark-content" />
+      <SafeAreaView>
+        <ScrollView
+          contentInsetAdjustmentBehavior="automatic"
+          style={styles.scrollView}>
+          <Header />
+          {global.HermesInternal == null ? null : (
+            <View style={styles.engine}>
+              <Text style={styles.footer}>Engine: Hermes</Text>
             </View>
-          </ScrollView>
-        </SafeAreaView>
-      </Fragment>
-    );
-  }
-}
+          )}
+          <View style={styles.body}>
+            <View testID="appRootView" style={styles.container}>
+              <View style={styles.header}>
+                <Text style={styles.text}>Example DateTime Picker</Text>
+              </View>
+              <View style={styles.button}>
+                <Button
+                  testID="datePickerButton"
+                  onPress={showDatepicker}
+                  title="Show date picker default!"
+                />
+              </View>
+              <View style={styles.button}>
+                <Button
+                  testID="datePickerButton"
+                  onPress={showDatepickerSpinner}
+                  title="Show date picker spinner!"
+                />
+              </View>
+              <View style={styles.button}>
+                <Button
+                  testID="timePickerButton"
+                  onPress={showTimepicker}
+                  title="Show time picker!"
+                />
+              </View>
+              <View style={styles.button}>
+                <Button
+                  testID="timePickerIntervalButton"
+                  onPress={showTimepickerWithInterval}
+                  title="Show time picker (with 5 min interval)!"
+                />
+              </View>
+              <View style={styles.header}>
+                <Text testID="dateTimeText" style={styles.dateTimeText}>
+                  {mode === 'time' && moment.utc(date).format('HH:mm')}
+                  {mode === 'date' && moment.utc(date).format('MM/DD/YYYY')}
+                </Text>
+              </View>
+              {show && (
+                <DateTimePicker
+                  minuteInterval={interval}
+                  testID="dateTimePicker"
+                  timeZoneOffsetInMinutes={0}
+                  value={date}
+                  mode={mode}
+                  is24Hour
+                  display={display}
+                  onChange={onChange}
+                  style={styles.iOsPicker}
+                />
+              )}
+            </View>
+          </View>
+        </ScrollView>
+      </SafeAreaView>
+    </>
+  );
+};
 
 const styles = StyleSheet.create({
   scrollView: {
@@ -142,4 +167,9 @@ const styles = StyleSheet.create({
     fontSize: 16,
     fontWeight: 'normal',
   },
+  iOsPicker: {
+    flex: 1,
+  },
 });
+
+export default App;
