@@ -7,22 +7,24 @@
 
 package com.reactcommunity.rndatetimepicker;
 
-import javax.annotation.Nullable;
-
-import java.util.Calendar;
-import java.util.Locale;
 import android.annotation.SuppressLint;
-
 import android.app.DatePickerDialog;
 import android.app.DatePickerDialog.OnDateSetListener;
 import android.app.Dialog;
 import android.content.Context;
 import android.content.DialogInterface;
 import android.content.DialogInterface.OnDismissListener;
+import android.content.DialogInterface.OnClickListener;
 import android.os.Build;
 import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
 import android.widget.DatePicker;
+
+import java.util.Calendar;
+import java.util.Locale;
 
 @SuppressLint("ValidFragment")
 public class RNDatePickerDialogFragment extends DialogFragment {
@@ -32,6 +34,8 @@ public class RNDatePickerDialogFragment extends DialogFragment {
   private OnDateSetListener mOnDateSetListener;
   @Nullable
   private OnDismissListener mOnDismissListener;
+  @Nullable
+  private static OnClickListener mOnNeutralButtonActionListener;
 
   @Override
   public Dialog onCreateDialog(Bundle savedInstanceState) {
@@ -45,15 +49,18 @@ public class RNDatePickerDialogFragment extends DialogFragment {
     instance.updateDate(date.year(), date.month(), date.day());
   }
 
-  static DatePickerDialog createDialog(Bundle args, Context activityContext, @Nullable OnDateSetListener onDateSetListener) {
-    final Calendar c = Calendar.getInstance();
+  static @NonNull
+  DatePickerDialog getDialog(
+          Bundle args,
+          Context activityContext,
+          @Nullable OnDateSetListener onDateSetListener) {
+
     final RNDate date = new RNDate(args);
     final int year = date.year();
     final int month = date.month();
     final int day = date.day();
 
     RNDatePickerDisplay display = RNDatePickerDisplay.DEFAULT;
-    DatePickerDialog dialog = null;
 
     if (args != null && args.getString(RNConstants.ARG_DISPLAY, null) != null) {
       display = RNDatePickerDisplay.valueOf(args.getString(RNConstants.ARG_DISPLAY).toUpperCase(Locale.US));
@@ -62,22 +69,31 @@ public class RNDatePickerDialogFragment extends DialogFragment {
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
       switch (display) {
         case CALENDAR:
-          dialog = new RNDismissableDatePickerDialog(activityContext,
-            activityContext.getResources().getIdentifier("CalendarDatePickerDialog", "style", activityContext.getPackageName()),
-            onDateSetListener, year, month, day);
-          break;
         case SPINNER:
-          dialog = new RNDismissableDatePickerDialog(activityContext,
-            activityContext.getResources().getIdentifier("SpinnerDatePickerDialog", "style", activityContext.getPackageName()),
-            onDateSetListener, year, month, day);
-          break;
-        case DEFAULT:
-          dialog = new RNDismissableDatePickerDialog(activityContext, onDateSetListener, year, month, day);
-          break;
+          String resourceName = display == RNDatePickerDisplay.CALENDAR
+                  ? "CalendarDatePickerDialog"
+                  : "SpinnerDatePickerDialog";
+          return new RNDismissableDatePickerDialog(
+                  activityContext,
+                  activityContext.getResources().getIdentifier(
+                          resourceName,
+                          "style",
+                          activityContext.getPackageName()),
+                  onDateSetListener,
+                  year,
+                  month,
+                  day);
+        default:
+          return new RNDismissableDatePickerDialog(
+                  activityContext,
+                  onDateSetListener,
+                  year,
+                  month,
+                  day
+          );
       }
     } else {
-      dialog = new RNDismissableDatePickerDialog(activityContext, onDateSetListener, year, month, day);
-
+      DatePickerDialog dialog = new RNDismissableDatePickerDialog(activityContext, onDateSetListener, year, month, day);
       switch (display) {
         case CALENDAR:
           dialog.getDatePicker().setCalendarViewShown(true);
@@ -87,6 +103,21 @@ public class RNDatePickerDialogFragment extends DialogFragment {
           dialog.getDatePicker().setCalendarViewShown(false);
           break;
       }
+      return dialog;
+    }
+  }
+
+  static DatePickerDialog createDialog(
+          Bundle args,
+          Context activityContext,
+          @Nullable OnDateSetListener onDateSetListener) {
+
+    final Calendar c = Calendar.getInstance();
+
+    DatePickerDialog dialog = getDialog(args, activityContext, onDateSetListener);
+
+    if (args != null && args.containsKey(RNConstants.ARG_NEUTRAL_BUTTON_LABEL)) {
+      dialog.setButton(DialogInterface.BUTTON_NEUTRAL, args.getString(RNConstants.ARG_NEUTRAL_BUTTON_LABEL), mOnNeutralButtonActionListener);
     }
 
     final DatePicker datePicker = dialog.getDatePicker();
@@ -133,5 +164,9 @@ public class RNDatePickerDialogFragment extends DialogFragment {
 
   /*package*/ void setOnDismissListener(@Nullable OnDismissListener onDismissListener) {
     mOnDismissListener = onDismissListener;
+  }
+
+  /*package*/ void setOnNeutralButtonActionListener(@Nullable OnClickListener onNeutralButtonActionListener) {
+    mOnNeutralButtonActionListener = onNeutralButtonActionListener;
   }
 }
