@@ -8,21 +8,19 @@
 
 package com.reactcommunity.rndatetimepicker;
 
-import android.app.Dialog;
+import java.text.DateFormat;
+import java.util.Locale;
+
 import android.app.TimePickerDialog;
 import android.app.TimePickerDialog.OnTimeSetListener;
-import android.content.Context;
 import android.content.DialogInterface;
+import android.content.DialogInterface.OnClickListener;
 import android.content.DialogInterface.OnDismissListener;
 import android.os.Build;
 import android.os.Bundle;
-import android.text.format.DateFormat;
-
+import androidx.annotation.Nullable;
 import androidx.fragment.app.DialogFragment;
-
-import java.util.Locale;
-
-import javax.annotation.Nullable;
+import javafx.scene.control.Dialog;
 
 @SuppressWarnings("ValidFragment")
 public class RNTimePickerDialogFragment extends DialogFragment {
@@ -32,6 +30,8 @@ public class RNTimePickerDialogFragment extends DialogFragment {
   private OnTimeSetListener mOnTimeSetListener;
   @Nullable
   private OnDismissListener mOnDismissListener;
+  @Nullable
+  private static OnClickListener mOnNeutralButtonActionListener;
 
   @Override
   public Dialog onCreateDialog(Bundle savedInstanceState) {
@@ -45,7 +45,11 @@ public class RNTimePickerDialogFragment extends DialogFragment {
     instance.updateTime(date.hour(), date.minute());
   }
 
-  static TimePickerDialog createDialog(Bundle args, Context activityContext, @Nullable OnTimeSetListener onTimeSetListener) {
+  static TimePickerDialog getDialog(
+          Bundle args,
+          Context activityContext,
+          @Nullable OnTimeSetListener onTimeSetListener) {
+
     final RNDate date = new RNDate(args);
     final int hour = date.hour();
     final int minute = date.minute();
@@ -70,36 +74,26 @@ public class RNTimePickerDialogFragment extends DialogFragment {
     }
 
     if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-      if (display == RNTimePickerDisplay.CLOCK) {
-        return new RNDismissableTimePickerDialog(
-          activityContext,
-          activityContext.getResources().getIdentifier(
-            "ClockTimePickerDialog",
-            "style",
-            activityContext.getPackageName()
-          ),
-          onTimeSetListener,
-          hour,
-          minute,
-          minuteInterval,
-          is24hour,
-          display
-        );
-      } else if (display == RNTimePickerDisplay.SPINNER) {
-        return new RNDismissableTimePickerDialog(
-          activityContext,
-          activityContext.getResources().getIdentifier(
-            "SpinnerTimePickerDialog",
-            "style",
-            activityContext.getPackageName()
-          ),
-          onTimeSetListener,
-          hour,
-          minute,
-          minuteInterval,
-          is24hour,
-          display
-        );
+      switch (display) {
+        case CLOCK:
+        case SPINNER:
+          String resourceName = display == RNTimePickerDisplay.CLOCK
+                  ? "ClockTimePickerDialog"
+                  : "SpinnerTimePickerDialog";
+          return new RNDismissableTimePickerDialog(
+                  activityContext,
+                  activityContext.getResources().getIdentifier(
+                          resourceName,
+                          "style",
+                          activityContext.getPackageName()
+                  ),
+                  onTimeSetListener,
+                  hour,
+                  minute,
+                  minuteInterval,
+                  is24hour,
+                  display
+          );
       }
     }
     return new RNDismissableTimePickerDialog(
@@ -111,6 +105,18 @@ public class RNTimePickerDialogFragment extends DialogFragment {
       is24hour,
       display
     );
+  }
+
+  static TimePickerDialog createDialog(
+          Bundle args, Context activityContext,
+          @Nullable OnTimeSetListener onTimeSetListener) {
+
+    TimePickerDialog dialog = getDialog(args, activityContext, onTimeSetListener);
+
+    if (args != null && args.containsKey(RNConstants.ARG_NEUTRAL_BUTTON_LABEL)) {
+      dialog.setButton(DialogInterface.BUTTON_NEUTRAL, args.getString(RNConstants.ARG_NEUTRAL_BUTTON_LABEL), mOnNeutralButtonActionListener);
+    }
+    return dialog;
   }
 
   @Override
@@ -127,5 +133,9 @@ public class RNTimePickerDialogFragment extends DialogFragment {
 
   public void setOnTimeSetListener(@Nullable OnTimeSetListener onTimeSetListener) {
     mOnTimeSetListener = onTimeSetListener;
+  }
+
+  /*package*/ void setOnNeutralButtonActionListener(@Nullable OnClickListener onNeutralButtonActionListener) {
+    mOnNeutralButtonActionListener = onNeutralButtonActionListener;
   }
 }
