@@ -1,8 +1,6 @@
 module.exports = {
   preset: 'react-native',
-  modulePathIgnorePatterns: [
-    '<rootDir>/example/node_modules',
-  ],
+  modulePathIgnorePatterns: ['<rootDir>/example/node_modules'],
   transform: {
     '^.+\\.js$': require.resolve('react-native/jest/preprocessor.js'),
   },
